Creating project structure.
