+++ src/containers/App/index.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import Finder from '../Finder';
+
+const App = () => (
+  <Finder />
+);
+
+export default App;
