+++ src/index.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import App from './containers/App';
+
+const Root = () => (
+  <App />
+);
+
+export default Root;
